+++ src/handlers/command-handler/typings.d.ts
@@ -1,4 +1,4 @@
-import { Client } from 'discord.js';
+import { ChatInputCommandInteraction, Client, ContextMenuCommandInteraction } from 'discord.js';
 import { ContextCommandObject, SlashCommandObject } from '../../../typings';
 
 export interface CommandHandlerOptions {
@@ -6,9 +6,20 @@ export interface CommandHandlerOptions {
     commandsPath: string;
     devGuildIds: string[];
     devUserIds: string[];
-    validations: Function[];
+    devRoleIds: string[];
+    customValidations: Function[];
+    skipBuiltInValidations: boolean;
 }
 
 export interface CommandHandlerData extends CommandHandlerOptions {
     commands: Array<SlashCommandObject | ContextCommandObject>;
+    builtInValidations: Array<BuiltInValidation>;
 }
+
+export interface BuiltInValidationParams {
+    targetCommand: SlashCommandObject | ContextCommandObject;
+    interaction: ChatInputCommandInteraction | ContextMenuCommandInteraction;
+    handlerData: CommandHandlerData;
+}
+
+export type BuiltInValidation = ({}: BuiltInValidationParams) => boolean | void;