+++ src/handlers/command-handler/validations/guildOnly.ts
@@ -1,0 +1,12 @@
+import { BuiltInValidationParams } from '../typings';
+
+module.exports = ({ interaction, targetCommand }: BuiltInValidationParams) => {
+    if (targetCommand.options?.guildOnly && !interaction.inGuild()) {
+        interaction.reply({
+            content: '❌ This command can only be used inside a server.',
+            ephemeral: true,
+        });
+
+        return true;
+    }
+};