+++ src/handlers/command-handler/validations/userPermissions.ts
@@ -1,0 +1,20 @@
+import { BuiltInValidationParams } from '../typings';
+
+module.exports = ({ interaction, targetCommand }: BuiltInValidationParams) => {
+    const memberPermissions = interaction.memberPermissions;
+
+    if (targetCommand.options?.userPermissions && memberPermissions) {
+        for (const permission of targetCommand.options.userPermissions) {
+            const hasPermission = memberPermissions.has(permission);
+
+            if (!hasPermission) {
+                interaction.reply({
+                    content: `❌ You do not have enough permission to run this command. Required permission: \`${permission}\``,
+                    ephemeral: true,
+                });
+
+                return true;
+            }
+        }
+    }
+};